+++ app/src/main/java/com/fastaccess/github/base/callback/UpdateTabCount.kt
@@ -6,5 +6,5 @@ import com.fastaccess.data.model.FragmentType
  * Created by Kosh on 16.10.18.
  */
 interface UpdateTabCount {
-    fun updateCount(type: FragmentType, count: Long)
+    fun updateCount(type: FragmentType, count: Int)
 }
