+++ data/src/main/java/com/fastaccess/data/model/CountModel.kt
@@ -2,4 +2,4 @@ package com.fastaccess.data.model
 
 import com.google.gson.annotations.SerializedName
 
-data class CountModel(@SerializedName("totalCount") var totalCount: Long? = null)
+data class CountModel(@SerializedName("totalCount") var totalCount: Int? = null)
